+++ src/test/java/org/drugis/addis/problems/service/MeasurementsServiceTest.java
@@ -83,7 +83,7 @@ public class MeasurementsServiceTest {
 
     Map<Long, String> trialverseVariables = new HashMap<>();
     trialverseVariables.put(outcomeId, "variable name");
-    when(triplestoreService.getTrialverseVariables(project.getTrialverseId(), analysis.getStudyId(), outcomeMap.keySet())).thenReturn(trialverseVariables);
+    when(triplestoreService.getTrialverseVariables(project.getTrialverseId().longValue(), analysis.getStudyId().longValue(), outcomeMap.keySet())).thenReturn(trialverseVariables);
 
     Measurement measurement = new Measurement((long) studyId, outcomeId, measurementMomentId, alternativeId, MeasurementAttribute.RATE, integerValue, null);
     ObjectNode jsonNode = mapper.convertValue(measurement, ObjectNode.class);
@@ -96,7 +96,7 @@ public class MeasurementsServiceTest {
 
     assertEquals(expected, measurements);
 
-    verify(triplestoreService).getTrialverseVariables(project.getTrialverseId(), analysis.getStudyId(), outcomeMap.keySet());
-    verify(triplestoreService).getTrialverseVariables(project.getTrialverseId(), analysis.getStudyId(), outcomeMap.keySet());
+    verify(triplestoreService).getTrialverseVariables(project.getTrialverseId().longValue(), analysis.getStudyId().longValue(), outcomeMap.keySet());
+    verify(triplestoreService).getTrialverseVariables(project.getTrialverseId().longValue(), analysis.getStudyId().longValue(), outcomeMap.keySet());
   }
 }
